Added box spawn script, updated RViz planning scene
The frames of the boxes when they were added manually in teh gazebo world were stuck at origin, so I used a spawn script to spawn the boxes on the table. I've also edited the box positions because some of the previous positions weren't reacheable by the robot, and updated the same in RViz.